requirements.txt
tensorflow_text
pandas
numpy
yfinance

matplotlib
scikit-learn
ta